requirements.txt
evaluate==0.1.2
numpy==1.21.2
opencv_python==4.5.4.60
Pillow==9.1.1
